+++ code/src/nuvla/job/actions/__init__.py
@@ -23,6 +23,7 @@ __all__ = [basename(f)[:-3] for f in modules if isfile(f) and not f.endswith('__
 
 
 class Actions(object):
+
     actions = {}
 
     @classmethod