+++ code/src/nuvla/job/actions/dummy_test_action.py
@@ -6,7 +6,8 @@ from ..actions import action
 
 @action('dummy_test_action')
 class DummyTestActionJob(object):
-    def __init__(self, executor, job):
+
+    def __init__(self, _, job):
         self.job = job
         self.api = job.api