+++ code/src/nuvla/job/actions/nuvlabox_decommission.py
@@ -7,18 +7,31 @@ from ..actions import action
 
 @action('decommission_nuvlabox')
 class NuvlaBoxDeleteJob(object):
-    def __init__(self, executor, job):
+
+    def __init__(self, _, job):
         self.job = job
         self.api = job.api
         self.error = 0
 
-    def delete_api_key(self, nuvlabox_id):
-        credentials = self.api.search('credential',
+    def delete_linked_resources(self, collection, nuvlabox_id):
+        try:
+            entries = self.api.search(collection,
                                       filter='parent="{}"'.format(nuvlabox_id),
                                       select='id').resources
+            for entry in entries:
+                entry_id = entry.id
+                try:
+                    self.api.delete(entry.id)
+                except Exception:
+                    self.error += 1
+                    logging.warning('problem deleting resource {}.'.format(entry_id))
+        except Exception:
+            # An exception when querying is probably caused by the collection
+            # not existing. Simply log and ignore this.
+            logging.warning('problem querying collection {}.'.format(collection))
 
-        for credential in credentials:
-            self.api.delete(credential.id)
+    def delete_api_key(self, nuvlabox_id):
+        self.delete_linked_resources('credential', nuvlabox_id)
 
     # FIXME: This will currently leave orphan data-object and data-record resources!
     def delete_s3_credential(self, credential_id):
@@ -87,17 +100,12 @@ class NuvlaBoxDeleteJob(object):
                 self.error += 1
                 logging.warning('problem deleting resource {}.'.format(infra_service_group_id))
 
+    def delete_peripherals(self, nuvlabox_id):
+        # If the nuvlabox-peripheral collection doesn't exist, then this acts as a no-op.
+        self.delete_linked_resources('nuvlabox-peripheral', nuvlabox_id)
+
     def delete_status(self, nuvlabox_id):
-        entries = self.api.search('nuvlabox-status',
-                                  filter='parent="{}"'.format(nuvlabox_id),
-                                  select='id').resources
-        for entry in entries:
-            entry_id = entry.id
-            try:
-                self.api.delete(entry.id)
-            except Exception:
-                self.error += 1
-                logging.warning('problem deleting resource {}.'.format(entry_id))
+        self.delete_linked_resources('nuvlabox-status', nuvlabox_id)
 
     def delete_nuvlabox(self):
         nuvlabox_id = self.job['target-resource']['href']
@@ -120,9 +128,13 @@ class NuvlaBoxDeleteJob(object):
 
             self.job.set_progress(30)
 
-            self.delete_api_key(nuvlabox_id)
+            self.delete_peripherals(nuvlabox_id)
 
             self.job.set_progress(40)
+
+            self.delete_api_key(nuvlabox_id)
+
+            self.job.set_progress(50)
 
             if self.error == 0:
                 try: