+++ code/src/nuvla/job/distributor.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+import sys
 import time
 import logging
 
@@ -9,19 +10,22 @@ from .base import Base
 class Distributor(Base):
     def __init__(self):
         super(Distributor, self).__init__()
+        self.collect_interval = None
 
     def _job_distributor(self):
         logging.info('I am {} and I have been elected to distribute "{}" jobs'
                      .format(self.name, self._get_jobs_type()))
-        while not self.stop_event.is_set():
+        while not Base.stop_event.is_set():
             for cimi_job in self.job_generator():
                 try:
                     logging.info('Distribute job: {}'.format(cimi_job))
                     self.api.add('job', cimi_job)
-                except:
-                    logging.error('Failed to distribute job: {}'.format(cimi_job))
+                except Exception:
+                    logging.error('Failed to distribute job: {}.'.format(cimi_job))
                     time.sleep(0.1)
+            time.sleep(self.collect_interval)
         logging.info('Distributor properly stopped.')
+        sys.exit(0)
 
     def _start_distribution(self):
         election = self._kz.Election('/election/{}'.format(self._get_jobs_type()), self.name)
@@ -31,8 +35,8 @@ class Distributor(Base):
 
     # ----- METHOD THAT CAN/SHOULD BE IMPLEMENTED IN DISTRIBUTOR SUBCLASS -----
     def job_generator(self):
-        """This is a generator function that produces a sequence of Job(s) to be added to the nuvla server.
-        This function must be overridden by the user subclass.
+        """This is a generator function that produces a sequence of Job(s) to be added to the nuvla
+        server. This function must be overridden by the user subclass.
         """
         raise NotImplementedError()